+++ mobile/android/android-components/buildSrc/src/main/java/Dependencies.kt
@@ -4,8 +4,8 @@
 
 // Synchronized version numbers for dependencies used by (some) modules
 private object Versions {
-    const val kotlin = "1.2.61"
-    const val coroutines = "0.23.4"
+    const val kotlin = "1.2.71"
+    const val coroutines = "0.30.2"
 
     const val junit = "4.12"
     const val robolectric = "3.8"